OddEvenCheck.java
This code is a basic simple code for find a number wheather it is an Odd number or Even number. 
In here, number variable has devided by 2 and if there is 1 as a reminder, it is an odd number while if there is no any ramainders, it can be considered as an Even number.